requirements.txt
flask
mcp
openai
pydantic-settings
python-dotenv
PyYAML
llama-index-core
llama-index-readers-file
llama-index-llms-openai-like
llama-index-embeddings-openai-like
pypdf
